@@ -16,18 +16,18 @@
     "@testing-library/user-event": "^13.2.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
+    "@types/react": "^18.0.29",
     "@types/react-chartjs-2": "^2.5.7",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^18.0.11",
     "@types/react-router-dom": "^5.3.3",
-    "axios": "^0.27.2",
+    "axios": "^1.3.4",
     "chart.js": "^3.8.0",
-    "clsx": "^1.1.1",
+    "clsx": "^1.2.1",
     "date-fns": "^2.28.0",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-chartjs-2": "^4.1.0",
-    "react-dom": "17.0.2",
-    "react-router-dom": "5.3.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.0",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
