@@ -1,3 +1,3 @@
-export default interface ParamTypes {
+export type ParamTypes = {
   id: string;
 }
